Update SqlAlchemy settings
Future versions of SqlAlchemy will not allow tracking of modifications of objects and emitting of signals. This version will issue a warning if this value is not explicitly set. This has a high memory overhead and the flask-sqlalchemy docs encourage you to turn it off. This change turns this feature off.